Excluded crafting result slot from quick move stack actions. Moved #quickMoveStack to terminal menu superclass.
